Install sip files directly in the .install file.
Rather than relying on the upstream build system to install them and
then copying.

This is needed to fix the race condition when doing the arch+indep
build: if the arch part installed the files earlier, the mv command
in the indep part will fail with “Directory not empty” error.